+++ examples/vue-consumer/src/main.ts
@@ -1,10 +1,6 @@
 import { createApp, onMounted, ref } from "vue";
-import {
-  registerComplexComponents,
-  sampleCustomers
-} from "@fahimc/react-web-component-bridge-test-components";
-
-registerComplexComponents();
+import "@fahimc/react-web-component-bridge-test-components/web-components";
+import { sampleCustomers } from "@fahimc/react-web-component-bridge-test-components";
 
 createApp({
   setup() {