+++ packages/test-components/src/web-components.ts
@@ -1,0 +1,5 @@
+import { registerExampleWebComponents } from "./index";
+
+registerExampleWebComponents();
+
+export * from "./index";